requirements.txt
opencv-python
Pillow
pytesseract
flask
flask_scss